deleting answer with ajax

+++ app/controllers/answers_controller.rb
@@ -14,12 +14,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author_of?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer successfully deleted.'
-    else
-      render 'questions/show'
-    end
+    answer.destroy if current_user.author_of?(answer)
   end
 
   private
